# Conflicts:
#	include/moonraker_advanced_api.h
#	src/api/moonraker_advanced_api.cpp
#	src/calibration/belt_tension_calibrator.cpp
#	src/ui/ui_panel_belt_tension.cpp
#	translations/de.yml
#	translations/en.yml
#	translations/es.yml
#	translations/fr.yml
#	translations/it.yml
#	translations/ja.yml
#	translations/pt.yml
#	translations/ru.yml
#	translations/zh.yml
#	ui_xml/panel_belt_tension.xml
#	ui_xml/translations/de.xml
#	ui_xml/translations/en.xml
#	ui_xml/translations/es.xml
#	ui_xml/translations/fr.xml
#	ui_xml/translations/it.xml
#	ui_xml/translations/ja.xml
#	ui_xml/translations/pt.xml
#	ui_xml/translations/ru.xml
#	ui_xml/translations/translations.xml
#	ui_xml/translations/zh.xml

+++ tests/unit/filament_consumption_tracker_test_access.h
@@ -43,7 +43,7 @@ struct FilamentConsumptionTrackerTestAccess {
     /// print_state_enum_ subject (useful when the test sets up fixtures before
     /// the aggregate observer fires).
     static void force_print_state(PrintJobState s) {
-        FilamentConsumptionTracker::instance().on_print_state_changed(static_cast<int>(s));
+        FilamentConsumptionTracker::instance().on_print_state_changed(s);
     }
 
     /// Invoke the per-extruder handler directly — lets routing tests verify
